+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - blanks between `;` separated paths are ignored
   - do **not** use double or single quotes around paths
 - added `$PYTHON` to context menu variables
+- added clear message after commit select box
 -